SemanticsProperties API renames and comments
This performs the following semantics property renames:
- hidden to invisibleToUser, for consistency with the visibleToUser
accessibility property we map it to, and with a comment saying it
should only be used in niche occlusion cases.
- All the range-related properties renamed for clarity and consistency
with existing accessibility properties.

Secondly, I deleted the comments on the SemanticsProperty object (except
for a @see) and did a first pass at improving the comments on the (more
commonly used) SemanticsPropertyReceiver property setters.  Maintaining
two copies of every comment was proving burdensome, and we should be
able to improve our API documentation quality faster in the future with
only one.

Test: Existing tests
Relnote: "SemanticsPropertyReceiver.hidden was renamed to invisibleToUser and marked @ExperimentalComposeUiApi.
AccessibilityRangeInfo was renamed to ProgressBarRangeInfo.
stateDescriptionRange was renamed to progressBarRangeInfo.
AccessibilityScrollState was renamed to ScrollAxisRange.
horizontalAccessibilityScrollState was renamed to horizontalScrollAxisRange.
verticalAccessibilityScrollState was renamed to verticalScrollAxisRange."

Change-Id: Id31487aa7cbddf4d3d163999afae458cdab4dc8a