Update formatting for day of week in Japanese.
Currently, we use a single character abbreviation for days of the week
in Japanese, e.g., '金', which literally means 'gold'. However, this
is not clear when displayed on the watch face and the full days of the
week take at most 3 characters, e.g., '金曜日'.

This commit adds an override for Japanese that prefers the full length
day of the week name.

Bug: 177041763
Test: ./gradlew :wear:wear-complications-provider:test
Change-Id: I33310f185ee67fecdb6d9ab575716d9c840a22be