Migrate Hilt Workers to use Dagger Assisted Inject
Update the API from @WorkerInject to @HiltWorker similar to the migration from @ViewModelInject to @HiltViewModel. The main difference is that workers use Dagger's assisted inject which means their constructor have to be annotated with @AssistedInject and both the Context and WorkerParameters with @Assisted. These changes makes these two API have some parity as they behave similarly.

Test: Hilt Worker Integration App
Relnote: Replace @WorkerInject with @HiltWorker. @HiltWorker is now a type annotation and requires the usage of @AssistedInject in the constructor.
Change-Id: Ic2f15c63880a02ed082e9205fcad7acdb2b38751