Allow anim resource values of 0 when navigating to an activity
When navigating to an Activity with extra NavOptions, if you set the
animation resource values to 0, the resources can not be found.

We should allow resource values to be set to 0 to prevent any Activity
animations.

RelNote: "Fixed regression caused by b/171364502 where navigating to an
activity with an animation resource value of 0 caused a
ResourceNotFoundException."
Test: tested in sample app from b/171364502
Bug: 176819931

Change-Id: I7aedb282e3bd4534758035fe17bfcdb3287be604