Rename CompositionLifecycleObserver to RememberObserver
This change effectively renames CompositionLifecycleObserver
to RememberObserver by introducing the new interface and
deprecating the old one.

Test: CompositionTests
Fixes: b/177330384
Relnote: """RememberObserver replaces CompositionLifecycleObserver
and CompositionLifecycleObserver is now deprecated.

`RememberObserver` is a replacement for
`CompositionLifecycleObserver` with modified semantics and
renamed methods. Changing to the new API can be done mechanically
for objects that are only remembered once which is, and continues to
be, the recommended practice. Howevever, if a reference was
remembered more than once in a composition `onRemembered` is called
for each reference where `onEnter` is only called once. `onEnter`
was called multiple time if the object was used in subcompositions,
such as  `WithConstraints` and `Scaffold` making the single
`onEnter` call guarantee unreliable and it was removed for
`RememberObserver`.

`RememberObserver` adds `onAbandoned` which is called if the
`RememberObserver` instance is returned from the callback passed
to `remember` but was not remembered in the composition state
and, therefore, will never have `onRemembered` called. This can
occur if an exception terminates composition before completing or
the composition is discarded because the state is was producing
a composition for is no longer current or otherwise is no longer
needed. If the instance of `RememberObserver` following the single
reference recommendation above is tracking an external resource
both `onForgotten` and `onAbandoned` each indicate that the
resource is no longer needed. If the object is tracking work
started or resources allocated in `onRemembered`,  `onAbandoned`
can be ignored as it will not be called if `onRemembered` is
called.
"""

Change-Id: I02c364f517507abd6a5c071fb527192ad1d77239