Move media2 modules into standardized dir names
Recently media2/media2-exoplayer was moved to media2/exoplayer to
consolidate with the other modules in its groupd, but the actual
convention in androidx modules is to be labeled group-module. While not
every group follows this convention, this change broke Copybara's config
due to misplacing an excluded symlink, and it would be better to fix
this properly than simply change Copybara to mirror the misconfigured
dirname.

Relnote: N/A
Test: ./gradlew bOS
Change-Id: I4f1a06499ad400cfb76574c77b84b79d3367a73a