Add clearAndSetSemantics behavior on same layout node
We realized that with the previous behavior (it used to only clear the
children), the majority of the uses of this would require an extra Box.
This changes it to also apply to the semantics it wraps on the same
node, so that it can be passed in as a modifier argument.

Test: SemanticsTests.clearAndSetSemanticsSameLayoutNode

Change-Id: I0a7dd55b1251712d2da93161d07feea964dc2684