Overhaul benchmark output - on by default in safe location
Fixes: 172376362
Test: ./gradlew bench:b-b:installRelease && adb shell am instrument -w -r -e debug false androidx.benchmark.benchmark.test/androidx.benchmark.junit4.AndroidBenchmarkRunner

Change default write dir to a location that shouldn't require
requestLegacyExternalStorage=true.

This allows us to enable output by default, as it's a necessary part
of debugging a macrobenchmark, and enables simpler setup for
microbenchmark CI / JSON usage as well.

Also, fixes default argument behavior.

Note: I'm still not able to locally reproduce the crash that we've
seen in CI when when requestLegacyExternalStorage is not set to true,
despite trying Q, R, and Q emulator. Punting actual removal of
requestLegacyExternalStorage flag into separate CL for this reason.

Change-Id: Id34103b2fb898a8cd0d90ebadc3c2db8551f23c1