Move cursor set and selection cancel to onTap
Note: this was reviewed on another branch before aosp/1510334

CoreTextField used a custom press/drag based on pressIndicator.
In the current state the press that this gesture detector catches
intervenes with the scroll press/drag. This causes the selection
to disappear while scrolling.

Moreover, there is a TODO in the drag gesture filter which
describes that it is not the proper gesture filter.

Finally, currently we actually handle tap rather than press, or
In other words the UX is the same.

We will need to write a custom press/drag handler to TextField
that can move the cursor as the finger moves. We have a ticket for
that.

Therefore this CL removes the custom drag handler, handles selection
and cursor offset in onTap.

RelNote: N/A

Test: Treehugger
Test: Manually tested on demo app
      Text > Input Fields > Scrollable text fields

Bug: 174208237
Change-Id: I7c9e8a313746a341a71cfdaa10bc45600d66bb1c