Add infra for testing generated code
This CL adds basic support for testing generated code. It is not as clever
as in java compile testing, specifically, it does not do structural diffing.
Instead, we simply compare trimmed lines while skipping empty lines.

It is not great but Room rarely does this kind of testing so it is good
enough for now.

I've also added an XRoundEnv for KSP and change its
getElementsAnnotatedWith API to return type elements. It is only used in
tests in Room and we don't need other elements. I'll try to remove root
elements as well as there are some tests using it w/o needing it.

Bug: 160322705
Test: GeneratedCodeMatchTest

Change-Id: Ie550a6edc73a4a9b8e08933f8a32481244146422