Estimate counting loaded items when all pages in presenter are dropped
This handles the rare scenario where pages are dropped so fast in
fetcher, that all pages in the presenter no longer exist in fetcher
and then it processes viewport hints before presenter receives any of
the drop events. In the prepend case, this normally leads to OOB due to
reading directly from loaded pages in fetcher, but in the append case
will just skip the count causing an incorrect number of items to be
loaded. As a workaround, this change simply clamps the count for pages
that would normally be associated with fulfilling prefetchDistance, and
estimates items to load using PagingConfig.pageSize. However once
presenter catches up with fetcher it should bind items and new viewport
hints to allow eventual consistency.

Test: ./gradlew paging:paging-common:test
Fixes: 178448265
Change-Id: I99ac230f0910cf1bb7232d9999d9767f052a6ecb