Read order of fields from the .class file
This CL fixes a bug in KSP where the order of fields returned in a
KSClassDeclaration does not necessarily match the order they were
declared in the class. It only impacts compiled kotlin classes because
Kotlin compiler orders field names when serializing the metadata.
https://github.com/google/ksp/issues/250

This particularly hurts Room as we generate the table structure based on
this order. Even though Room is clever enough to consider the schema the
same, the effects are still visible in both schema files and generated
sql database.

This CL is a port of a possible implementation in KSP:
https://github.com/google/ksp/pull/260
Unfortunately, merging that CL would mean KSP doing a lot more work for
each class. The plan there is to first fix it in the kotlin compiler if
possible and if not, add an API to get sorted fields (to avoid the cost
when not necessary). Meanwhile, this CL is a band aid solution in Room
so that developers won't have unexpected schema file changes when they
move to KSP from KAPT.

As this solution is best effort, it doesn't fail compilation if it fails
to read fields. I've added a strict mode to XProcessing so that we can
run it in tests with a requirement to always find the order but on
developer's machine, it won't require this. This strict mode can be
turned on via a system property, which we can use to collect more bug
information when it doesn't work.

Fixes: 177471948
Test: OrderOfFieldsTest
Change-Id: If0cf346a7f92dbbf2094e0076ea9ed02de4e32fc