Add fallback reporting for elements w/o location
When we try to report an error on an element that is coming from
dependencies, its location cannot be found, making the error report less
understandable.
In javac, we were detecting this case and using element utils to report
the error. As we need something similar for KSP, I've moved this
reporting to XProcessing where each XElement instance implements a
property to report its location.

Bug: 177475650
Test: FallbackLocationInformationTest
Change-Id: I05a2d4098e14667192321a336105a0f7def441f7