Cancel infinite animations in tests
Infinite animations are now started through withInfiniteAnimation, which
checks the InfiniteAnimationPolicy and applies it if present when
withFrameNanos or withFrameMillis is called.

ComposeTestRule installs an InfiniteAnimationPolicy in the
applyCoroutineContext that cancels the coroutine when that policy is
applied.

This enables us to remove special handling of demos with infinite
animations in DemoTest, and we can now remove blinkingCursorEnabled.

Fixes: 151940543
Bug: 168695905
Test: ./gradlew compose:animation:animation-core:test && \
        ./gradlew compose:animation:animation-core:cC && \
        ./gradlew compose:ui:ui-test:cC
Relnote: "Introduced an `InfiniteAnimationPolicy` coroutine context
element that will be applied in infinite animations. By default no
policy is installed, except when running tests with `ComposeTestRule`."

Change-Id: I50ec421d7db495459a61c9282dbc2bfbc1f1ad02