Fix postsubmit failure of uiwidgets.rotations.*Test
For b/172808769, b/172808699 and other related failed tests, the uiwidgets.rotations.Image*Test will fail on Pixel 1 easily because of heavily opening/closing camera and capture-session.

From the failed log, it is observed that there is always a extra capture request before closing capture session/camera. After tracing the code, it is triggerred by cacheInteropConfig() because the camera state is active.

The fix makes the mCameraInternal.detachUseCases invoked first and it will make no usecase attach the camera and the state is inactive. Hence, the cacheInteropConfig() will not update session, and then reduce the capture-request update for the capture-session.

After checking, the test fail rate decreased a lot for the fix. For the fix, there is no hurt on Camera2 logic and no impact on other tests and it can improve the efficiency when camera is closed.

Bug: b/172808769, b/172808699
Test: manual test on Pixel, Pixel 2 and Pixel 3 for uiwidgets.rotations.*Test
&& ./gradlew camera:integration-tests:camera-testapp-uiwidgets:connectedCheck

Change-Id: Iaa513777c83741a117a43b2c2923bd5c9ff35aca