Do not continuously attempt to force open the camera when a camera slot is open
With aosp/1555358, CameraX attempts to force open the camera whenever a use case is attached/multiple use cases are attached to the camera, or whenever a camera slot becomes available. The second scenario causes the camera to continuously attempt to (force) open the camera when the camera is not available, e.g. on API level 21 when the camera is already opened by another camera client. This is because of the following flow:

-> Initial attempt to force open the camera
-> CameraRegistry reports that there is a camera slot available: Proceed to open the camera
-> Camera open attempt fails with error ERROR_CAMERA_IN_USE
-> Schedule a camera reopen attempt
-> Scheduled reopen attempt runs, the camera is still not available, so the state is set to PENDING_OPEN.
-> CameraRegistry sees that there is a camera in a PENDIN_OPEN state and there is an open camera slot, it notifies the camera by triggering onOpenAvailable().
-> An attempt to force open the camera is made again, thus repeating this entire flow.

This CL reverts back to simply attempting to open the camera, until a better way to force open the camera in this scenario is implemented and tested.

Bug: 179275971
Test: Manual testing
Change-Id: Ief4ff80bc9184c793f878ad246d9596798626a1f