Simplifies Indication APIs
Previously IndicationInstance was responsible for both writing state / launching animations and then reading state to draw indication. Typically we want to model this as a one-way pipeline, so IndicationInstance has now been simplified to be only responsible for drawing indication.

Indication#createInstance() has been changed to rememberUpdatedInstance(InteractionState) to convey these responsibilities more clearly - Indication is responsible for observing InteractionState and creating an IndicationInstance which can draw visual effects for the current state.

This CL leaves the current Ripple implementation as is; it will be refactored in a future CL.

Bug: b/152525426
Test: updateApi
Test: existing tests
Relnote: "Changed Indication#createIndication() to Indication#rememberUpdatedIndication(InteractionState) and removes InteractionState parameter from IndicationInstance#drawIndication(). IndicationInstance should only be responsible for drawing visual effects, and not launching animations / writing state in response to InteractionState changes. These animations and state writes should happen within `rememberUpdatedIndication()` instead. The `indication` parameter in `Modifier.indication` was also changed to be a required parameter."
Change-Id: Ic1764c0417b25cd0a0dbb96ed3e1b0974618c4ca