Defer item insertion changes to post layout
This CL fixes a bug in RecyclerView where it was applying position
changes for additions into the ViewHolders in pre-layout for cached items.
This means such changes will be visible in pre-layout, breaking
consistency.

It was already postponing them when applying to existing children so I
think this is a relatively safe change. Similarly, changes for
deletions, moves are also postponed except for removals in non-visible
items (there is no reason to re-layout them).

Bug: 179181037
Test: GridLayoutManagerTest
Change-Id: I0330caeb57c3a6e5b6f9805047a122ad13da4253