Add getPackageToDatabases.
Returns a mapping of package to database names within that package. This
will eventually be used by VisibilityStore to create uris for each
visibility document. VisibilityStore currently calls getPrefixesLocked
and uses the prefixes as uris, but the concept of "prefix" is an
internal implementation detail of AppSearchImpl. This will prevent
leakage of "prefixes" outside of AppSearchImpl.

This change intentionally does not remove getPrefixesLocked so that
exporting to framework will be smoother. Once platform
VisibilityStore has migrated away from getPrefixesLocked, we can safely
remove it from upstream here.

Bug: 180058203
Test: ./gradlew appsearch:appsearch:connectedCheck --info --daemon
  ./gradlew appsearch:appsearch-local-storage:connectedCheck --info
  --daemon
Change-Id: I93ab0dfea6bf19ae1f0237e068b6fa22c899fab6