Kotlin Compile Testing optimizations
This CL implements two optimizations for KCT.

a) Right now, it resolves the host classpath for each run, taking 50-60
ms for each compilation (and we run ~3 of them per test). I've added
caching for common jars that it resolves from host classpath so that it
does not need to resolve it for each test.

b) When we inherit classpath (which we always do), the size of the
classpath seems to have a signficant impact on kotlin compilation time.
Instead, this CL picks a filtered view from the whole host classpath.
(see linked issue for measurements).

After these changes, TableEntityProcessorTest (which runs ~400
compilations) takes ~4:30 min instead of ~8 min on my machine.

https://github.com/tschuchortdev/kotlin-compile-testing/issues/113

Bug: 160322705
Test: existing test.

Change-Id: I8a738c76879d44a31bb22b050466a2dc49ef45e1