Create AutoMigrationWriter and associated tests.
The AutoMigrationWriter will be called by the DatabaseProcessingStep and will take in an AutoMigrationResult containing all schema changes detected. The AutoMigrationResult will be unpacked and each schema change detected will be processed by a separate function that will add the necessary statements into the migrate() function that will be generated. This CL implements functionality for the simple schema change of adding a new column.

Test: AutoMigrationWriterTest.kt
Bug: 74441348
Change-Id: I0e38fda93e2556b92d1fd5bdb07be8dd8bae23e1