Complete transitions if the container is not laid out
If the container has not be laid out, beginDelayedTransition() is a
no-op and since the transition is never started, it never ends so the
SpecialEffectController operation never gets completed.

In the case where the container has not been laid out, we should complete
the operation immediately.

Unlike transitions, Animations run whether they have been laided out or
not and they should continue to do so.

RelNote: "Child fragments with transitions will now properly reach
RESUMED"
Test: enterExitChildTransitions
Bug: 180825150

Change-Id: Ic11e655bd4ed206485c2762a1af4e5bc7aa4cf9c