Run LazyListScrollingBenchmark tests only starting with Q
Previously the benchmarks caught when we started doing a bit more work in Modifier.background(). But the reality is  it only happens because we run benchmark without RenderNodes support. In the real app we will not redraw items when we move them, so such changes intentionally shouldn't affect our tests. According to AndroidComposeTestCaseRunner we can use RenderNode implementations on Q+ and from M to O. I tried to do if checks right inside the test but then the benchmark rule complaints that the test didn't start any benchmark in such conditions. So the only way to restrict the test run I found is @SdkSuppress(minSdkVersion = Build.VERSION_CODES.Q)

Test: manually
Fixes: 179010059
Change-Id: I36d1c3f7216b0684451323f7968088dc5f0f36de