Add Activity.setPipAnimationHintView extension
Allows setting an android.view.View as the reference to be used
for animating PIP entry and exit transitions. The view's
position is tracked between onStart and onStop and is used to
update the PictureInPictureParams.Builder.setSourceRectHint
as the view's position changes.

Bug: 167447693
Test: run PipActivity in integration test app
Relnote: "Added new API Activity#setPipAnimationHintView to update
the PipParams' source rect hint whenever the view moves."

Change-Id: I9063de7990e997607b60384ea3f37970b902c39a