Ensure Fragments do not clear view transitionNames
When transitioning views are nested within ViewGroups that have
transitionNames but are not transitioning groups, the views are captured
as sharedElement views multiple times within the DefaultSpecialEffects
controller. As a result of this, as we collect and clear the transition
names, we capture a null name and end up restoring the null name to the
view once the shared element transition has finished.

We should only capture each shared element view once and therefore we
only restore it once, with the proper transition name.

RelNote: "Views within in a shared element hierarchy will no longer have
their transition name cleared when doing a shared element transition."
Test: testNestedSharedElementView
Bug: 179934757

Change-Id: I4d4a6d7770d5c6d54e4c647559d5cabae71f0051