Implementing SQL statements needed for performing complex schema changes.
This CL implements generating the migrate() function that accounts for all schema changes except for table/column renames/deletes and changes involving FTS tables. The follow-up CL will account for these remaining changes.

In addition, changes have been made in RoomDatabase to provide the user the ability to get a particular generated autoMigration in the form of a Migration object, by specifying the "from" and "to" versions of the desired autoMigration. This is a workaround so that users can test their autoMigrations using the MigrationTestHelper API's runMigrationsAndValidate() function. If the user requests an autoMigration that has not been generated, RoomDatabase will throw an IllegalArgumentException.

Test: The AutoMigrationTest.kt migrates the database from version 1 to version 2. The schemas at these two different version involve a wide variety of schema changes performed, ranging from a simple change as adding a new table to changing an existing index.
Bug: 180395129
Change-Id: Ia1f3ca7db5debf664a0791e8a3f93d9ea68bc966