Introduce navigation-compose-material and support for bottom sheet navigation destinations
For easier integration with compose-material's ModalBottomSheetLayout, this library provides convenience wrappers and a BottomSheetNavigator to be able to declare and navigate to bottom sheet destinations, similar to bottom sheet dialog destinations in vanilla navigation.
BottomSheetNavigator drives a ModalBottomSheet's sheetContent and state, controlling visibility according to the back stack and composing and disposing destinations that are navigated to or away from. BottomSheetNavigator offers the sheetContent as a composable that should be set as the ModalBottomSheetLayout's sheetContent.

- Composable bottom sheet destinations can be added using NavGraphBuilder#bottomSheet
- The BottomSheetNavigator needs to be registered with the NavController's navigatorProvider
- Provides a convenience wrapper around ModalBottomSheetLayout that accepts a BottomSheetNavigator
- Extracted BackStackEntryIdViewModel and the SaveableStateHolder.SaveableStateProvider extension into their own file in navigation-compose because they are needed in navigation-compose-material

Test: Added tests for new behavior (BottomSheetNavigatorTest and SheetContentHostTest), ran with ./gradlew test connectedCheck. More tests to be added.
Demos: Introduced BottomSheetNavDemo
Fixes: b/180247978

Change-Id: I51d2c86ce01ef13d68622d368ae1efaee51ffd7c