Add scroll to item actions for LazyLists in tests
Defines a set of SemanticsPropertyKeys that are added in LazyList and
used in test actions to the effect of scrolling a LazyList to an item
that might not yet be composed.

Bug: 182259407
Bug: 178483889
Bug: 161584524
Fix: 183714443
Test: LazySemanticsTest, ScrollToIndexTest, ScrollToKeyTest
Relnote: "Added `SemanticsActions.ScrollToIndex` to scroll a list with
indexed items to the item with a certain index, and
`SemanticsProperties.IndexForKey` to get the index of an item in a list
with keyed items. Both actions are implemented by LazyList."
Relnote: "Added `SemanticsNodeInteraction.performScrollToIndex` that
scrolls a list to the given index, and
`SemanticsNodeInteraction.performScrollToKey` that scrolls a list to the
item with the given key."

Change-Id: I4fe63399fb620794651e1973730658877bcfeff4