Remove docs on invisible component in AlignmentLine.
"merger" is an internal val, and AlignmentLine is a sealed class with no public constructor
This means that there is nothing to document, so docs generation assumes you made a typo in "merger".
The same documentation already exists on all subclasses of sealed class AlignmentLine, so the docs are sufficient without the @param on AlignmentLine.

Test: docs-only CL

Change-Id: I498ad77dafe15b8b7cdcb623b3ecf867fc138fd8