Implementing support for handling all schema changes except FTS Table related ones.
This CL implements functionality that is able to detect, categorize and generate a migrate() function for all complex and simple schema changes (except FTS Table related changes), including renames and deletes at the column and table level.

We now are able to get input from the user via the repeatable annotations: @RenameColumn, @RenameTable, @DeleteColumn, @DeleteTable. In the cases where the schema change classification fails in case of an ambiguity (e.g. a table is missing in the new version, no new tables have been discovered, and the table has not been annotated with @RenameTable or @DeleteTable), an error message is displayed to the user with sample code using these annotations.

In addition, we are now able to provide the autoMigrations to the database only by a list of @AutoMigration annotations. When the user opts to have a custom AutoMigrationCallback, they can include this interface in the annotation. It is important to note that to use the repeatable change annotations, the user must have a custom AutoMigrationCallback interface, as the repeatable annotations are only used on this callback. If the user does not provide a callback implementation, and potential column/table rename/deletes are encountered during the AutoMigration generation process, an error message will be displayed with sample code usage, notifying the user that they are required to define a custom callback and use the annotations to clarify the change.

Lastly, changes have been made in this CL to handle the case where the user includes an AutoMigration annotation with the same from and to version values as one of their custom Migration definitions. In this case, we will avoid overriding the pre-existing user defined Migration, and will not generate an autoMigration.

Test: The AutoMigrationTest.kt
Bug: 180395129, 183434667, 183007590, 182251019
Change-Id: Ic834c5d75753b5279e6d6c54745474258f816991