Only run ChannelLogoUtilsTest on TV devices
This test requires the TvProvider to be present in order to pass. On TV devices, the test passes as expected. However, on phones (and mobile emulators) the tests fail since the content provider is not available on those form factors.

Since the CI process runs on a phone emulator, false falures were reported which is why the test was originally supressed. After further investigation, we can enable the tests to run on any form factor but the actualy assertion for logic only occurs when the TvProvider is present.

Test: ChannelLogoUtilsTest (both on TV and Phone emulator)
Bug: 70905391
Change-Id: Ie72b7c9174ad37ae773c1fc2a15baf5c9957a85f