Add routes to NavDestination
This change introduces the concept of a route to NavDestination. A route
is String that defines the path to the Destination. Setting the route on
a NavDestination will replace the any id that was set previously  and
adds a deep link to the destination. You can set a route on the
NavDestination programatically, or through xml.

RelNote: "You can now add a route to your NavDestination"
Test: NavDestinationTest
Bug: 172823546

Change-Id: I77daf563f480ae54b9728a78e7696ea96d19f264