Run FieldProcessorTest with KSP
I had to modify tests a bit due to flexible types in kotlin.

When Kotlin compiler parses a java source without nullability
annotations, its type will be flexible betwee (T..T?) which means the
type equality in the test does not work.
This should not affect caching (which is why Field is a data class) but
affects the test. I've added nullability annotations whenever possible.

For boxed arrays (e.g. Integer[]), there is no way to fix the
nullability of type argument in Kotlin Array (Array<Int..Int?>). For
that test, I just changed it to assert type via typenames.

Also, the test that detects unbound generics do not apply to KSP as they
are always bound to Any..Any?. Instead, i changed the test to assert
that compilation failed with "cannot save this type" error.

Bug: 160322705
Test: FieldProcessorTest
Change-Id: I9f622b7aaff7d7b5f8050836044fb142f828d811