Ensure FragmentContainerView dispatches insets once
FragmentContainerView overrides onApplyWindowInsets() and dispatches new
insets to all children, but since it is a ViewGroup, and does not
consume the window insets, it makes a call to dispatchApplyWindowInsets
again. On this 2nd pass, if the first child consumes the WindowInsets,
all other children receive empty insets.

We should have FragmentContainerView override dispatchApplyWindowInsets
and dispatch them to the child views there.

RelNote: "When using `FragmentContainerView`, a new set of
`WindowInsets` are now dispatched to child views only once."
Test: modified windowInsetsDispatchToChildren test
Bug: 172153900

Change-Id: I63f685e6715334a77e477180ae94771eeef827c3