Fix popEnter animations when popping setPrimaryNavFragment
When doing a FragmentTransactions, if you setCustomAnimations with a
popEnter animation, do an op that triggers an animation (i.e. add, replace, etc),
and then do setPrimaryNavFragment, when you pop that transaction, there will be
no popEnter animation. The reason for this is the setPrimaryNavFragment
Op clears all of the animations for the incoming fragment.

We should only change animations for operations that actually trigger
them.

RelNote: "PopEnter animations will now be present when popping a
setPrimaryNavFragment operation."
Test: FragmentAnimationTest
Bug: 183877426

Change-Id: I38c877ae34c825b3a451ab325b6f683de00a2457